Moved CourseSelectCard/ under CourseSelectColumn/
Updated imports where necessary
