+++ autoscheduler/frontend/src/components/CourseSelectColumn/CourseSelectCard/CollapsedCourseCard/CollapsedCourseCard.css
