+++ autoscheduler/frontend/src/components/CourseSelectColumn/CourseSelectCard/CourseSelectCard.tsx
