+++ autoscheduler/frontend/src/components/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/BasicSelect/BasicSelect.tsx
@@ -3,8 +3,8 @@ import {
   Checkbox, List, ListItem, ListItemIcon, ListItemText, FormLabel,
 } from '@material-ui/core';
 import { useSelector, useDispatch } from 'react-redux';
-import { RootState } from '../../../../redux/reducer';
-import { updateCourseCard } from '../../../../redux/actions/courseCards';
+import { RootState } from '../../../../../redux/reducer';
+import { updateCourseCard } from '../../../../../redux/actions/courseCards';
 
 interface BasicSelectProps {
   id: number;
