+++ autoscheduler/frontend/src/components/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/SectionSelect/SectionSelect.tsx
@@ -3,12 +3,12 @@ import {
   List, ListItemText, ListItem, Checkbox, ListItemIcon, Typography, ListSubheader,
 } from '@material-ui/core';
 import { useSelector, useDispatch } from 'react-redux';
-import Meeting, { MeetingType } from '../../../../types/Meeting';
-import { SectionSelected } from '../../../../types/CourseCardOptions';
-import { RootState } from '../../../../redux/reducer';
+import Meeting, { MeetingType } from '../../../../../types/Meeting';
+import { formatTime } from '../../../../../timeUtil';
+import { SectionSelected } from '../../../../../types/CourseCardOptions';
+import { RootState } from '../../../../../redux/reducer';
 import * as styles from './SectionSelect.css';
-import { updateCourseCard } from '../../../../redux/actions/courseCards';
-import { formatTime } from '../../../../timeUtil';
+import { updateCourseCard } from '../../../../../redux/actions/courseCards';
 
 interface SectionSelectProps {
   id: number;
